♻️ refactor: DesktopNavBar & NavBar 통합
